CLEAN: getattr(value) on any data returns the python value (read only) of the data
