+++ src/SofaPython3/DataHelper.cpp
@@ -401,138 +401,4 @@ py::array getPythonArrayFor(BaseData* d)
     return memcache[d];
 }
 
-
-/// Make a python version of our BaseData.
-/// If possible the data is exposed as a numpy.array to minmize copy and data conversion
-py::object toPython(BaseData* d, bool writeable)
-{
-    const AbstractTypeInfo& nfo{ *(d->getValueTypeInfo()) };
-    /// In case the data is a container with a simple layout
-    /// we can expose the field as a numpy.array (no copy)
-    if(nfo.Container() && nfo.SimpleLayout())
-    {
-        if(!writeable)
-        {
-            getPythonArrayFor(d);
-            return PythonFactory::toPython(d);
-        }
-        return getPythonArrayFor(d);
-    }
-
-    /// If this is not the case we return the converted datas (copy)
-    return convertToPython(d);
-}
-
-
-template<class SrcType>
-void copyFromListOf(const AbstractTypeInfo& nfo, void* ptr, size_t index, py::object o);
-
-template<> void copyFromListOf<double>(const AbstractTypeInfo& nfo, void* ptr, size_t index, py::object o)
-{
-    nfo.setScalarValue(ptr, index, py::cast<double>(o));
-}
-
-template<> void copyFromListOf<int>(const AbstractTypeInfo& nfo, void* ptr, size_t index, py::object o)
-{
-    nfo.setIntegerValue(ptr, index, py::cast<int>(o));
-}
-
-template<> void copyFromListOf<std::string>(const AbstractTypeInfo& nfo, void* ptr, size_t index, py::object o)
-{
-    nfo.setTextValue(ptr, index, py::cast<std::string>(o));
-}
-
-
-template<class DestType>
-void copyFromListOf(BaseData& d, const AbstractTypeInfo& nfo, const py::list& l)
-{
-    /// Check if the data is a single dimmension or not.
-    py::buffer_info dstinfo = toBufferInfo(d);
-
-    if(dstinfo.ndim>2)
-        throw py::index_error("Invalid number of dimension only 1 or 2 dimensions are supported).");
-
-    if(dstinfo.ndim==1)
-    {
-        void* ptr = d.beginEditVoidPtr();
-        if( (size_t)dstinfo.shape[0] != l.size())
-            nfo.setSize(ptr, l.size());
-
-        for(size_t i=0;i<l.size();++i)
-        {
-            copyFromListOf<DestType>(nfo, ptr, i, l[i]);
-        }
-        d.endEditVoidPtr();
-        return;
-    }
-    void* ptr = d.beginEditVoidPtr();
-    auto t=getShape(&d);
-    if( (size_t)dstinfo.shape[0] != l.size())
-    {
-        if( !nfo.setSize(ptr, l.size()*nfo.size()) )
-            throw std::runtime_error("Unable to resize vector to match list size. Is the data type resizable ?");
-
-        /// Update the buffer info entry to take into account the change of size.
-        dstinfo = toBufferInfo(d);
-    }
-
-    for(auto i=0;i<dstinfo.shape[0];++i)
-    {
-        py::list ll = l[i];
-        for(auto j=0;j<dstinfo.shape[1];++j)
-        {
-            copyFromListOf<DestType>(nfo, ptr, i*dstinfo.shape[1]+j, ll[j]);
-        }
-    }
-
-    d.endEditVoidPtr();
-    return;
-}
-
-template<>
-void copyFromListOf<std::string>(BaseData& d, const AbstractTypeInfo& nfo, const py::list& l)
-{
-    void* ptr = d.beginEditVoidPtr();
-    if( (size_t)nfo.size() != l.size())
-        nfo.setSize(ptr, l.size());
-
-    for(size_t i=0;i<l.size();++i)
-    {
-        copyFromListOf<std::string>(nfo, ptr, i, l[i]);
-    }
-    d.endEditVoidPtr();
-    return;
-}
-
-
-void fromPython(BaseData* d, const py::object& o)
-{
-    const AbstractTypeInfo& nfo{ *(d->getValueTypeInfo()) };
-
-    if(!nfo.Container())
-    {
-        scoped_writeonly_access guard(d);
-        if(nfo.Integer())
-            nfo.setIntegerValue(guard.ptr, 0, py::cast<int>(o));
-        else if(nfo.Text())
-            nfo.setTextValue(guard.ptr, 0, py::cast<py::str>(o));
-        else if(nfo.Scalar())
-            nfo.setScalarValue(guard.ptr, 0, py::cast<double>(o));
-        return ;
-    }
-
-    if(nfo.Integer())
-        return copyFromListOf<int>(*d, nfo, o);
-
-    if(nfo.Text())
-        return copyFromListOf<std::string>(*d, nfo, o);
-
-    if(nfo.Scalar())
-        return copyFromListOf<double>(*d, nfo, o);
-
-    std::stringstream s;
-    s<< "binding problem, trying to set value for "
-     << d->getName() << ", " << py::cast<std::string>(py::str(o));
-    throw std::runtime_error(s.str());
-}
 }
